Use the "npm start" way to run the app on Heroku instead of Procfile
Procfile syntax is different than Javascript or JSON and may lead to confusion
for students around meaningful whitespace and the weird Procfile extension-less
filename which is not consistent with the rest of the files in the app.

Heroku will use "npm start" over Procfile if that script is defined within the
"package.json" file during the application startup phase of deployment.