Used ImageOptim to compress and strip meta-data from common images
Many of these images have had their sizes reduced dramatically, up to 80% by
stripping the meta-data, compressing, and running through several lossless or
lossy algorithms. The result are images which are much less heavy and still look
good on the site. Mobile users will thank us.

Addtional compression is needed for the other images in the site, but these are
the most common images at this time. I'll figure out some kind of automatic
compression in the future to run as a task before deploy.