Add status badges, logo, and license links to the README intro
It's now easy to see if there are issues with the boot camp curriculum based on
the status of badges on the home pages and the services which serve the status
of those badges.

The logo for Burlington Code Academy is now prominent as well as the licenses
which will apply for the content and the code respectively.