Fix the build of review apps by using ENV vars in Heroku
The new version of review apps does not allow for environment variables to be
inherited by the application on build from either staging or production. All
environment variables have to be set and passed in from the Heroku UI or set
within the 'app.json' configuration file. To keep the configuration consistent,
we are using the Heroku UI to set the variables required in review.