Change the titilize solution to use a loop instead of .map
The .map higher order function has not yet been introduced in the class
curriculum and this can cause confusion when a student sees it for the
first time within a solution to a challenge. It feels like a trick that
they have not yet been taught. Instead we should show how to use a while
loop, which is a concept they should be familiar with at this point.