Upgrade to Heroku-22 and Ruby v3.1.4
Older versions of Heroku stacks are no longer supported, upgrade to allow for
maintenance and updates to the application. Since Heroku-18 was deprecated on
April 30th, 2023 no deploys to Heroku would be allowed.

Ruby 3.x and Heroku-22 will allow the application to be maintained for several
more years if needed.